% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ethnicityCpGs}
\alias{ethnicityCpGs}
\title{CpGs to predict ethnicity}
\format{
A character vector of length 1860
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31399127/}
}
\usage{
ethnicityCpGs
}
\description{
1860 CpGs used to predict ethnicity.

See \href{https://pubmed.ncbi.nlm.nih.gov/31399127/}{Yuan et al. 2019} for
details.
}
\keyword{datasets}
