% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nbeta}
\alias{nbeta}
\title{Coefficients to predict ethnicity.}
\format{
A [list] with three matrices corresponding to the three ethnicity 
classes
}
\usage{
nbeta
}
\description{
Coefficients to predict ethnicity.
}
\keyword{internal}
