% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
lhs
}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\examples{

c(1, 2, 3) \%>\% sum()
}
\keyword{internal}
