% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.r
\docType{data}
\name{pl_betas}
\alias{pl_betas}
\title{6 DNA methylation profiles from preeclampsia and healthy control placentas}
\format{A matrix}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196}
}
\usage{
pl_betas
}
\description{
"Genome wide DNA methylation profiling of normal and preeclampsia placental samples. Illumina
Infinium HumanMethylation450 BeadChip (450K array) was used to obtain DNA methylation profiles
in placental samples. Samples included 16 samples from healthy uncomplicated pregnancies and 8
samples from pregnancies affected by preeclampsia."
}
\details{
The DNA methylation data for 24 placental samples were downloaded from
[GSE75196](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196). After normalizing using
`minfi::preprocessNoob` and `wateRmelon::BMIQ`, the betas were filtered to 10,000 random CpGs +
those CpGs used in the gestational age clock and ethnicity classifier.

Reference: Yeung KR, Chiu CL, Pidsley R, Makris A et al. DNA methylation profiles in preeclampsia
and healthy control placentas. Am J Physiol Heart Circ Physiol 2016 May 15;310(10):H1295-303.
PMID: 26968548
}
\keyword{datasets}
