% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.r
\docType{data}
\name{pl_cell_cpgs_first}
\alias{pl_cell_cpgs_first}
\title{First trimester coefficients for placental cellular deconvolution from Yuan V et al. 2020.}
\format{
A [tibble][tibble::tibble-package] with coefficients for Trophoblasts, Stromal,
Endothelial, Hofbauer cells, nRBCs, and Syncytiotrophoblasts.
}
\usage{
pl_cell_cpgs_first
}
\description{
Reference:
PMID:
}
\keyword{datasets}
