% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pl_colors}
\alias{pl_colors}
\title{pl_colors}
\format{
An object of class \code{character} of length 6.
}
\usage{
pl_colors
}
\description{
A color palette for cell types. Used in Yuan V et al. 2020.
Contains colors for:

- Syncytiotrophoblast
- Trophoblast
- Stromal
- Hofbauer
- Endothelial
- nRBCs
}
\keyword{datasets}
