% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pl_ethnicity_features}
\alias{pl_ethnicity_features}
\title{Features from 450k/850k data used to predict ethnicity}
\format{A character vector of length 1860}
\source{
\url{https://github.com/wvictor14/Ethnicity_Inference_450k}
}
\usage{
pl_ethnicity_features
}
\description{
The features on the 450k/850k DNAm microarry used to predict ethnicity.
}
\keyword{datasets}
