% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_infer_ethnicity.R
\name{pl_infer_ethnicity}
\alias{pl_infer_ethnicity}
\title{Infers ethnicity using DNA methylation microarray data}
\usage{
pl_infer_ethnicity(betas, threshold = 0.75)
}
\arguments{
\item{betas}{n x m dataframe of methylation values on the beta scale (0, 1), where the variables
are arranged in rows, and samples in columns. Should contain all 1860 predictors.}

\item{threshold}{A probability threshold ranging from (0, 1) to call samples 'ambiguous'.
Defaults to 0.75.}
}
\value{
A m x 7 dataframe of predicted ethnicity information and associated probabilities.
}
\description{
\code{pl_infer_ethnicity} Uses a glmnet model to predict ethnicity using DNA methylation data.
}
\details{
A glmnet model fit using the package caret is applied to the data. The input data should
contain all 1860 predictors of the final model.
}
\examples{
## To predict ethnicity on 450k/850k samples

# Load placenta DNAm data
data(pl_rgset)

b <- getBeta(pl_rgset)     # 450k/850k methylation data
s <- getSnpBeta(pl_rgset)  # 450k/850k SNP data

testDat <- rbind(b, s)

pl_infer_ethnicity(testDat)

}
