% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pl_pDat}
\alias{pl_pDat}
\title{Sex, disease, and gestational age information associated with 
`pl_betas`. 

Taken from 
[GSE75196](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196).}
\format{
A [tibble][tibble::tibble-package]
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196}
}
\usage{
pl_pDat
}
\description{
Sex, disease, and gestational age information associated with 
`pl_betas`. 

Taken from 
[GSE75196](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196).
}
\keyword{datasets}
