% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet-deprecated.R
\name{planet-deprecated}
\alias{planet-deprecated}
\title{Deprecated functions in helloworld}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{pl_infer_ethnicity}}: This function is deprecated, and 
 will be removed in the next version of this package.
 \item \code{\link{pl_infer_age}}: This function is deprecated, and 
 will be removed in the next version of this package.
}
}
