% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet-package.R
\docType{package}
\name{planet-package}
\alias{planet}
\alias{planet-package}
\title{planet: Placenta elastic net DNA methylation ethnicity classifier}
\description{
This package contains a single function: a glmnet model which was developed on 450k DNA
  methylation data (but is compatible with 850k data as well) to infer ethnicity in placental 
  samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/wvictor14/planet}
  \item Report bugs at \url{http://github.com/wvictor14/planet}
}

}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

}
\keyword{internal}
