\name{PlanetScopeNICFI}
\alias{PlanetScopeNICFI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Planet Scope NICFI
}
\description{
Show a data frame with the avalilable images in the Planet Scope Tropical Normalized Analytic Biannual Archive and Planet Scope Tropical Normalizad Analytic Monthly Monitoring}
\usage{
  PlanetScopeNICFI()
}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{Name}{Text to name the folder for download images}
%  \item{DirPlanet}{Directory to planet API install}
%  \item{AOI}{Directory to Area of Interest Vector (any format sf can read ex: .GPKG, .shp)}
%  \item{Year}{2020 by default}
%  \item{Month}{9 for september, the only image available to the date }
%}

\value{Data frames with available images}

\author{Klaus Wolfgang Wiese Acosta}


\examples{
#Execute
PlanetScopeNICFI()
}

