% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkD3.R
\docType{package}
\name{planetvis-package}
\alias{planetvis-package}
\alias{planetvis}
\title{Tools for Creating D3 Network Graphs from R}
\description{
Creates D3 JavaScript network, tree, dendrogram, and Sankey graphs from R.
}
