% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plCellCpGsFirst}
\alias{plCellCpGsFirst}
\title{First trimester placental cell type coefficients}
\format{
A \link{matrix} with coefficients for Trophoblasts, Stromal,
Endothelial, Hofbauer cells, nRBCs, and Syncytiotrophoblasts.
}
\usage{
data(plCellCpGsFirst)
}
\description{
First trimester coefficients for placental cellular
deconvolution from \url{Yuan V et al. 2020}.

Reference: to be edited
PMID: to be edited
}
\keyword{datasets}
