% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plCellCpGsThird}
\alias{plCellCpGsThird}
\title{Third trimester placental cell type coefficients}
\format{
A \link{matrix} with coefficients for Trophoblasts, Stromal,
Endothelial, Hofbauer cells, nRBCs, and Syncytiotrophoblasts.
}
\usage{
data(plCellCpGsThird)
}
\description{
Third trimester coefficients for placental cellular
deconvolution from \url{Yuan V et al. 2020}.

Reference: to be edited
PMID: to be edited
}
\keyword{datasets}
