% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plPhenoData}
\alias{plPhenoData}
\title{Sample information accompanying \code{pl_betas}}
\format{
A \link[tibble:tibble-package]{tibble}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196}
}
\usage{
plPhenoData
}
\description{
Sex, disease, and gestational age information associated with
\code{pl_betas}.

Downloaded from the GEO accession:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196}{GSE75196}
}

Reference: Yeung KR, Chiu CL, Pidsley R, Makris A et al. DNA methylation
profiles in preeclampsia and healthy control placentas. Am J Physiol Heart
Circ Physiol 2016 May 15;310(10):H1295-303.
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75196}{PMID: 26968548}
}
\keyword{datasets}
