% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pl_ethnicity_features}
\alias{pl_ethnicity_features}
\title{Features from 450k/850k data used to predict ethnicity

The features on the 450k/850k DNAm microarry used to predict ethnicity.}
\format{
A character vector of length 1860
}
\source{
\url{https://epigeneticsandchromatin.biomedcentral.com/
articles/10.1186/s13072-019-0296-3}
}
\usage{
pl_ethnicity_features
}
\description{
Features from 450k/850k data used to predict ethnicity

The features on the 450k/850k DNAm microarry used to predict ethnicity.
}
\keyword{datasets}
