% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_infer_age.R
\name{pl_infer_age}
\alias{pl_infer_age}
\title{Predicts gestational age using DNA methylation microarray data}
\usage{
pl_infer_age(betas, type = "RPC")
}
\arguments{
\item{betas}{An n by m dataframe of methylation values on the beta scale (0,
1), where the CpGs are arranged in rows, and samples in columns. Should
contain all CpGs used in each clock}

\item{type}{One of the following: "RPC" (Robust), "CPC", (Control) or "RRPC"
(Refined Robust).}
}
\value{
A vector of length `m`, containing inferred gestational age.
}
\description{
\code{pl_infer_age} Multiplies the coefficients from one of three epigenetic
gestational age clocks by the corresponding CpGs in a supplied betas
`data.frame`.
}
\details{
Predicts gestational age using one of 3 placental gestational age
clocks: RPC, CPC, or refined RPC. Requires placental DNA methylation measured
on the Infinium 27K/450k/EPIC methylation array. If CpGs are missing, a
warning is returned.
}
\examples{

# Load placenta DNAm data
library(dplyr)
data(pl_betas)
data(pl_pDat)

pl_pDat \%>\%
  mutate(inferred_ga = pl_infer_age(pl_betas, type = "RPC"))
}
