% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet-package.R
\docType{package}
\name{planet-package}
\alias{planet}
\alias{planet-package}
\title{planet: Placental DNA methylation analysis tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains R functions to predict biological variables to from placnetal DNA methylation data generated from infinium arrays. This includes inferring ethnicity/ancestry, gestational age, and cell composition from placental DNA methylation array (450k/850k) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://victor.rbind.io/planet}
  \item \url{http://github.com/wvictor14/planet}
  \item Report bugs at \url{http://github.com/wvictor14/planet/issues}
}

}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

Authors:
\itemize{
  \item Wendy P. Robinson [contributor]
  \item Icíar Fernández-Boyano \email{iciarfernandez@outlook.com} [contributor]
}

}
\keyword{internal}
