% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet-package.R
\docType{package}
\name{planet-package}
\alias{planet}
\alias{planet-package}
\title{planet: Placenta elastic net DNA methylation ethnicity classifier}
\description{
This package contains R functions to infer ethnicity/ancestry, gestational age, and cell composition from placental DNA methylation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://victor.rbind.io/planet}
  \item Report bugs at \url{http://github.com/wvictor14/planet/issues}
}

}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

}
\keyword{internal}
