% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet-package.R
\docType{package}
\name{planet-package}
\alias{planet}
\alias{planet-package}
\title{planet: Placental DNA methylation analysis tools}
\description{
This package contains R functions to infer additional biological
    variables to supplemental DNA methylation analysis of placental data. This 
    includes inferring ethnicity/ancestry, gestational age, and cell composition
    from placental DNA methylation array (450k/850k) data. The package comes 
    with an example processed placental dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://victor.rbind.io/planet}
  \item \url{http://github.com/wvictor14/planet}
  \item Report bugs at \url{http://github.com/wvictor14/planet/issues}
}

}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

Other contributors:
\itemize{
  \item Wendy P. Robinson [contributor]
}

}
\keyword{internal}
