\name{PlanetScopeNICFI}
\alias{PlanetScopeNICFI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Planet Scope NICFI
}
\description{
Show a data frame with the avalilable images in the Planet Scope Tropical Normalized Analytic Biannual Archive and Planet Scope Tropical Normalizad Analytic Monthly Monitoring}
\usage{
  PlanetScopeNICFI()
}

\value{Data frames with available images}

\author{Klaus Wolfgang Wiese Acosta}


\examples{
#Execute
PlanetScopeNICFI()
}

