% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plColors}
\alias{plColors}
\title{A color palette for placental cell types}
\format{
An object of class \code{character} of length 6.
}
\usage{
plColors
}
\description{
A nice color palette for placental cell types.

Used in \url{Yuan V et al. 2020}.

Contains colors for:
\itemize{
\item Syncytiotrophoblast
\item Trophoblast
\item Stromal
\item Hofbauer
\item Endothelial
\item nRBCs
}
}
\keyword{datasets}
